Add files via upload
The objective of crop yield prediction using machine learning is to develop models 
that can accurately predict the yield of a crop given various environmental and 
farming factors. Agriculture plays a critical role in the global economy. With the 
continuing expansion of the human population, understanding worldwide crop 
yield is central to addressing food security challenges and reducing the impacts of 
climate change. Crop yield prediction is one of the challenging tasks in agriculture. 
It plays an essential role in decision making at global, regional, and field levels. 
Agricultural yield primarily depends on weather conditions (rain, temperature, 
etc), pesticides and accurate information about the history of crop yield is an 
important thing for making decisions related to agricultural risk management and 
future predictions. Machine learning is an important decision support tool for 
crop yield prediction, including supporting decisions on what crops to grow and 
what to do during the growing season of the crops. Several machine learning 
algorithms can be applied to support crop yield prediction.